Initial commit of issue 2064 for OSX
