--- iris/tests/bookmark/edit_this_bookmark_panel_for_bookmarked_page.py
@@ -1,40 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this file,
-# You can obtain one at http://mozilla.org/MPL/2.0/.
-
-
-from iris.test_case import *
-
-
-class Test(BaseTest):
-
-    def __init__(self):
-        BaseTest.__init__(self)
-        self.meta = '"Edit this bookmark"panel appears after a page is already bookmarked'
-        self.test_case_id = '163399'
-        self.test_suite_id = '2525'
-        self.locales = ['en-US']
-
-    def run(self):
-        soap_wiki_tab_pattern = Pattern('soap_wiki_tab.png')
-
-        navigate(LocalWeb.SOAP_WIKI_TEST_SITE)
-
-        soap_wiki_opened = exists(soap_wiki_tab_pattern, DEFAULT_SITE_LOAD_TIMEOUT)
-        assert_true(self, soap_wiki_opened, 'Soap wiki page opened')
-
-        bookmark_page()
-
-        stardialog_displayed = exists(Bookmarks.StarDialog.DONE, DEFAULT_FIREFOX_TIMEOUT)
-        assert_true(self, stardialog_displayed, 'Bookmark page dialog displayed')
-
-        click(Bookmarks.StarDialog.DONE)
-
-        restore_firefox_focus()
-
-        click(LocationBar.STAR_BUTTON_STARRED)
-
-        edit_stardialog_displayed = exists(Bookmarks.StarDialog.EDIT_THIS_BOOKMARK, DEFAULT_FIREFOX_TIMEOUT)
-        assert_true(self, edit_stardialog_displayed, 'The Edit This Bookmark popup is displayed under the star-shaped '
-                                                     'icon.')
-
